requirements.txt
astroid==1.4.8
blinker==1.4
click==6.6
contextlib2==0.5.4
expects==0.8.0
Flask==0.11.1
future==0.15.2
isort==4.2.5
itsdangerous==0.24
Jinja2==2.7.3
lazy-object-proxy==1.2.2
markdown2==2.2.1
MarkupSafe==0.23
mccabe==0.5.2
nose==1.3.7
pylint==1.6.4
python-dateutil==2.6.0
PyYAML==3.11
raven==5.32.0
six==1.10.0
Werkzeug==0.11.11
wrapt==1.10.8
